requirements.txt
vyper==0.3.1
eth-brownie==1.18.1
